# docs/api/rag-service.openapi.yaml
openapi: 3.0.3
info:
  title: AI Interview RAG Service API
  version: 0.2.0
  description: 独立 RAG 服务与 C++/Qt AI 面试客户端之间的 HTTP 契约。
servers:
  - url: http://127.0.0.1:8000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/knowledge-bases/{knowledge_base_id}/documents:
    parameters:
      - $ref: '#/components/parameters/KnowledgeBaseId'
      - $ref: '#/components/parameters/RequestIdHeader'
    post:
      tags: [Documents]
      summary: 上传文档并异步处理
      operationId: uploadDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                display_name:
                  type: string
                  minLength: 1
                  maxLength: 255
      responses:
        '202':
          description: 文档已接收并进入处理队列
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
    get:
      tags: [Documents]
      summary: 分页列出知识库文档
      operationId: listDocuments
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Cursor'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/DocumentStatus'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/knowledge-bases/{knowledge_base_id}/documents/{document_id}:
    parameters:
      - $ref: '#/components/parameters/KnowledgeBaseId'
      - $ref: '#/components/parameters/DocumentId'
      - $ref: '#/components/parameters/RequestIdHeader'
    get:
      tags: [Documents]
      summary: 获取文档详情和处理状态
      operationId: getDocument
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Documents]
      summary: 删除文档、文档块和向量
      operationId: deleteDocument
      responses:
        '204':
          description: 删除成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
  parameters:
    RequestIdHeader:
      name: X-Request-ID
      in: header
      required: false
      description: 客户端提供的链路追踪 ID
      schema:
        type: string
        minLength: 1
        maxLength: 128
    KnowledgeBaseId:
      name: knowledge_base_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
    DocumentId:
      name: document_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        maxLength: 512
  responses:
    BadRequest:
      description: 请求字段或参数无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 缺少或无效的认证信息
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 请求的资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 资源名称、内容或状态冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    PayloadTooLarge:
      description: 上传文件超过服务端限制
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnsupportedMediaType:
      description: 不支持的文件类型
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    DocumentStatus:
      type: string
      enum: [queued, processing, ready, failed]
    Document:
      type: object
      required:
        - id
        - knowledge_base_id
        - file_name
        - status
        - chunk_count
        - error_message
        - created_at
        - updated_at
      properties:
        id:
          type: string
        knowledge_base_id:
          type: string
        file_name:
          type: string
        status:
          $ref: '#/components/schemas/DocumentStatus'
        chunk_count:
          type: integer
          minimum: 0
        error_message:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DocumentResponse:
      type: object
      required: [request_id, document]
      properties:
        request_id:
          type: string
        document:
          $ref: '#/components/schemas/Document'
    DocumentPage:
      type: object
      required: [request_id, items, total]
      properties:
        request_id:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        next_cursor:
          type: string
          nullable: true
        total:
          type: integer
          minimum: 0
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: KNOWLEDGE_BASE_NOT_FOUND
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [request_id, error]
      properties:
        request_id:
          type: string
        error:
          $ref: '#/components/schemas/Error'